refactor(platform): convert provider dialogs to sidepanels
Replace FormDialog-based provider add/edit modals with Sheet-based
sidepanels matching the integration panel pattern. Extract inline
setup guide markup into a reusable CollapsibleGuide component.
Add edit-form validation and clarify read-only provider name field.
