+++ examples/integrations/ai-image/config.json
@@ -74,5 +74,5 @@
   "metadata": {
     "modelCapabilities": "Configured model: {{model}}\nModel capabilities:\n- gpt-image-1: sizes [1024x1024, 1024x1536, 1536x1024, auto], supports edit\n- dall-e-3: sizes [1024x1024, 1024x1792, 1792x1024], does NOT support edit\n- dall-e-2: sizes [256x256, 512x512, 1024x1024], supports edit\nIMPORTANT: Only use sizes supported by the configured model. Invalid sizes will cause an error."
   },
-  "setupGuide": "### Configuration Guide\n\n1. **domain** — Your API provider's base URL:\n   - OpenAI: `https://api.openai.com`\n   - Together AI: `https://api.together.xyz`\n\n2. **accessToken** — Your API key from the provider\n\n3. **model** — The image generation model ID:\n   - `gpt-image-1` — OpenAI's latest image model (sizes: 1024x1024, 1024x1536, 1536x1024, auto)\n   - `dall-e-3` — DALL-E 3 (sizes: 1024x1024, 1024x1792, 1792x1024, no edit support)\n   - `dall-e-2` — DALL-E 2 (sizes: 256x256, 512x512, 1024x1024)"
+  "setupGuide": "1. **domain** — Your API provider's base URL:\n   - OpenAI: `https://api.openai.com`\n   - Together AI: `https://api.together.xyz`\n\n2. **accessToken** — Your API key from the provider\n\n3. **model** — The image generation model ID:\n   - `gpt-image-1` — OpenAI's latest image model (sizes: 1024x1024, 1024x1536, 1536x1024, auto)\n   - `dall-e-3` — DALL-E 3 (sizes: 1024x1024, 1024x1792, 1792x1024, no edit support)\n   - `dall-e-2` — DALL-E 2 (sizes: 256x256, 512x512, 1024x1024)"
 }
