+++ services/platform/messages/de.json
@@ -1478,6 +1478,7 @@
       "name": "Anbietername",
       "namePlaceholder": "z. B. openai",
       "nameHelp": "Nur Kleinbuchstaben, Zahlen und Bindestriche.",
+      "nameReadonlyHelp": "Der Anbietername ist eine permanente Kennung und kann nicht geändert werden. Verwende den Anzeigenamen, um die Darstellung anzupassen.",
       "namePatternError": "Muss mit einem Buchstaben beginnen und darf nur Kleinbuchstaben, Zahlen und Bindestriche enthalten.",
       "displayNamePlaceholder": "z. B. OpenAI",
       "descriptionPlaceholder": "z. B. OpenAI GPT-Modelle",
@@ -1523,6 +1524,7 @@
       "providerActions": "Anbieteraktionen",
       "saveChanges": "Änderungen speichern",
       "addModelShort": "+ Modell hinzufügen",
+      "byomGuidanceTitle": "Erste Schritte",
       "byomGuidance": "Jeder OpenAI-kompatible Endpunkt funktioniert hier. Setze die Basis-URL auf deinen selbst gehosteten Inferenzserver (z. B. vLLM, Ollama, LocalAI) oder ein Gateway wie OpenRouter, um Modelle wie LLaMA, Mistral, Gemini und mehr zu nutzen.",
       "tagsRequired": "Mindestens ein Tag ist erforderlich",
       "modelsRequired": "Mindestens ein Modell ist erforderlich",
